Started code for random block spawning.
Changed Group.cs to exclude the game over function and remove the left and right movement of the blocks (kept the quicker drop key and rotate key).

Changed Playfield.cs to increase the vertical height for now, got an error since the game has a valid grid check and the size was too small for height climbed.

Changed Spawner.cs to have the character reference that will now spawn blocks 14 units above the character at random locations within (2, 9). Our grid is from 1 to 10 and at least one of our blocks has a protrusion on both sides so we must not spawn it at the barrier or else we get invalid placements.