Updated Group, Playfield, and Spawner files.
Group.cs was updated to include a blockFallTime variable that can be edited later on to increase the fall speed of blocks. Added mouse interactions that are commented out for now (was used to up the timescale). Added Q key for rotate and S key for quicker drop.

Increased the playfield width by 10 and height by 1800. The width is now 20 and height is 2000. Game is much less performant when height is increased to 2 million, probably because we are doing 2 million block checks each time the block falls a grid space.

Updated Spawner.cs to have fallTime and lastTime variables. Fall time is the amount we want to take away from total block fall time, and last time is used to track 100 seconds going by which will then lower the 1s block fall by 0.1 per every 100 seconds. Added a switch statement to catch blocks so they do not spawn outside the bounds of the walls.