Updated Scene, ClimbingCharacterController, Group, and Spawner.
Scene changes are anchoring the next block panel and changing its color.

Character script changes are massive and now has the added functionality for a mid-air block kick. Also changed the camera to only follow the characters Y position (we can change this back since the UI jittering was fixed).

Group was changed to now have a Set enum which stands for when the block is no longer active. This was needed to choose which block to kick. Also added the move left and move right as public functions that we can use through the character script.

Spawner was updated to move the next block as a child of the UI panel so there is no jitter. The block is later on removed from that parent and is its own block again.