Many changes and additions.
Edited the scene for the UI changes.

Edited character controller script to have block punch (rotate).

Edited group script to have a function for rotate instead of keypresses, and added a simple score system.

Added Score.cs to control the score text to be the value we set inside Group.cs.

Edited Spawner.cs to change the block falling speed (since it was somewhat requested) to starting value of 0.7f and decreases by 0.1f every 60 seconds.

Added Timer.cs to control the playtime timer texts so we can see how long we have been playing for.

Edited the pixel art font so we don't see the extra lines on the tops or bottoms, also added a space and underline character to get rid of warnings and errors.