Major edits and additions. Mostly UI related
Scene changed due to added elements to the UI.

Added ButtonHandler that controls how we interact with the 3 bottom right buttons.

Changed ClimbingCharacterController to add header for the skill cooldowns.

Changed code inside Group so it will not allow the player to drop blocks when the game is paused.

Changed Score to have height/top height included as a high score variant/substitute.

Added Skills which controls the visual aspect of when we can and can't use our block punch/kick/freeze time skills.

Changed Spawner to use a transform instead of game object to try to fix where the next blocks spawn inside the UI.