gameplay balances
Edited scene to change button interaction to mouse only.

Changed climbing character to have cam update if player y is too high, or time based.

Changed group to allow for more blocks to spawn, fixed the ghost/fake blocks on grid(Roxy), and added score fixes.

Playfield has been edited to 200, we want to shift the field with character to cut down on size(to be implemented).

Spawner was changed to spawn blocks off the camera's position, along with block fall speed tweaks.